test:update junit of metadata. (#340)
