+++ spring-cloud-starter-tencent-metadata-transfer/src/test/java/com/tencent/cloud/metadata/core/EncodeTransferMetadataZuulFilterTest.java
@@ -16,29 +16,27 @@
  *
  */
 
-package com.tencent.cloud.metadata.core.filter;
+package com.tencent.cloud.metadata.core;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
 
+import com.netflix.zuul.context.RequestContext;
 import com.tencent.cloud.common.constant.MetadataConstant;
 import com.tencent.cloud.common.util.JacksonUtils;
-import com.tencent.cloud.metadata.core.EncodeTransferMedataScgFilter;
 import org.assertj.core.api.Assertions;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.gateway.filter.GatewayFilterChain;
 import org.springframework.context.ApplicationContext;
-import org.springframework.mock.http.server.reactive.MockServerHttpRequest;
-import org.springframework.mock.web.server.MockServerWebExchange;
+import org.springframework.mock.web.MockMultipartHttpServletRequest;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
@@ -48,24 +46,31 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = RANDOM_PORT,
-		classes = EncodeTransferMedataScgFilterTest.TestApplication.class,
+		classes = EncodeTransferMetadataZuulFilterTest.TestApplication.class,
 		properties = { "spring.config.location = classpath:application-test.yml", "spring.main.web-application-type = reactive" })
-public class EncodeTransferMedataScgFilterTest {
+
+public class EncodeTransferMetadataZuulFilterTest {
 
 	@Autowired
 	private ApplicationContext applicationContext;
 
-	@Mock
-	private GatewayFilterChain chain;
+	private MockMultipartHttpServletRequest request = new MockMultipartHttpServletRequest();
+
+	@Before
+	public void init() {
+		RequestContext ctx = RequestContext.getCurrentContext();
+		ctx.clear();
+		ctx.setRequest(this.request);
+	}
 
 	@Test
-	public void testTransitiveMetadataFromApplicationConfig() throws UnsupportedEncodingException {
-		EncodeTransferMedataScgFilter filter = applicationContext.getBean(EncodeTransferMedataScgFilter.class);
-		MockServerHttpRequest.BaseBuilder<?> builder = MockServerHttpRequest.get("");
-		MockServerWebExchange exchange = MockServerWebExchange.from(builder);
-		filter.filter(exchange, chain);
-		String metadataStr = exchange.getRequest().getHeaders().getFirst(MetadataConstant.HeaderName.CUSTOM_METADATA);
-		String decode = URLDecoder.decode(metadataStr, StandardCharsets.UTF_8.name());
+	public void multiplePartNamesWithMultipleParts() throws UnsupportedEncodingException {
+		EncodeTransferMetadataZuulFilter filter = applicationContext.getBean(EncodeTransferMetadataZuulFilter.class);
+		filter.run();
+		final RequestContext ctx = RequestContext.getCurrentContext();
+		Map<String, String> zuulRequestHeaders = ctx.getZuulRequestHeaders();
+		String metaData = zuulRequestHeaders.get(MetadataConstant.HeaderName.CUSTOM_METADATA.toLowerCase());
+		String decode = URLDecoder.decode(metaData, StandardCharsets.UTF_8.name());
 		Map<String, String> transitiveMap = JacksonUtils.deserialize2Map(decode);
 		Assertions.assertThat(transitiveMap.size()).isEqualTo(1);
 		Assert.assertEquals(transitiveMap.get("b"), "2");
